add useEffect for fetching data from Firebase
