+++ src/components/firebase/firebaseConfig.js
@@ -1,7 +1,7 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth"
-import { collection, getFirestore, getDocs, doc, addDoc } from 'firebase/firestore'
+import { collection, getFirestore } from 'firebase/firestore'
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -20,23 +20,4 @@ const firebaseConfig = {
 export const app = initializeApp(firebaseConfig);
 export const auth = getAuth(app)
 export const database = getFirestore(app) //es necesario poner (app) ?
-
-
-// LUEGO LO MOVEMOS EN OTRO SITIO!!
-
-// collection ref
-const colRef = collection(database, 'movies') // colRef es collection reference
-
-let movies = [];  // eso devuelve un array con todas las pelis
-// get collection data
-getDocs(colRef)
-  .then((snapshot) => {
-    //console.log('snapshot:::', snapshot)
-    snapshot.docs.forEach((doc) => {
-      movies.push({ ...doc.data() })
-    })
-  })
-
-console.log('database: ', database)
-console.log('colRef: ', colRef)
-console.log('Movies: ', movies);
+export const moviesRef = collection(database, 'movies')  // moviesRef es collection reference
